Add first implementation of doman.
This is the first commit to doman. It contains 3 files, colors.sh,
manager.sh and README.org.

1. colors.sh - Utility file that defines unicodes for different colors
that I plan to use in doman.

2. manager.sh - The main file. This file parses a configuration file
and sets up symlinks for each of the target. It currently just creates
links only if there are no existing links or files, and let's the user
know in any other case